requirements.txt
fastapi
uvicorn
python-jose
passlib
python-multipart
bcrypt
python-dotenv
pymongo
dnspython